# Make `Device.set_current()` faster