--- a/cuda_core/cuda/core/experimental/_device.py
+++ b/cuda_core/cuda/core/experimental/_device.py
@@ -1202,13 +1202,13 @@ class Device:
             if int(ctx) == 0:
                 # use primary ctx
                 ctx = self._get_primary_context()
-                handle_return(driver.cuCtxPushCurrent(ctx))
+                handle_return(driver.cuCtxSetCurrent(ctx))
             else:
                 ctx_id = handle_return(driver.cuCtxGetDevice())
                 if ctx_id != self._id:
                     # use primary ctx
                     ctx = self._get_primary_context()
-                    handle_return(driver.cuCtxPushCurrent(ctx))
+                    handle_return(driver.cuCtxSetCurrent(ctx))
                 else:
                     # no-op, a valid context already exists and is set current
                     pass